+++ app/repository/video.go
@@ -6,8 +6,8 @@ import (
 	"time"
 )
 
-func PageVideoListByTime(lastTime time.Time, count int) []Video {
+func PageVideoListByTime(lastTime time.Time, count int64) []Video {
 	var videos []Video
-	config.Db.Table("videos").Where("publish_time < ?", lastTime).Order("publish_time desc").Limit(count).Find(&videos)
+	config.Db.Table("videos").Where("publish_time < ?", lastTime).Order("publish_time desc").Limit(int(count)).Find(&videos)
 	return videos
 }
