+++ app/schema/Follow.go
@@ -1,8 +1,8 @@
 package schema
 
 type Follow struct {
-	Id         int    `json:"id"`
-	UserId     int    `json:"user_id"`
-	FollowerId int    `json:"follower_id"`
+	Id         int64  `json:"id"`
+	UserId     int64  `json:"user_id"`
+	FollowerId int64  `json:"follower_id"`
 	ActionType string `json:"action_type"`
 }
