+++ app/schema/user.go
@@ -1,7 +1,7 @@
 package schema
 
 type User struct {
-	Id       int    `json:"id,omitempty"`
+	Id       int64  `json:"id,omitempty"`
 	Username string `json:"username"`
 	Password string `json:"password"`
 }
