+++ app/schema/video.go
@@ -3,8 +3,8 @@ package schema
 import "time"
 
 type Video struct {
-	Id          int    `json:"id"`
-	AuthorId    int    `json:"author_id"`
+	Id          int64  `json:"id"`
+	AuthorId    int64  `json:"author_id"`
 	PlayUrl     string `json:"play_url"`
 	CoverUrl    string `json:"cover_url"`
 	PublishTime time.Time
